+++ src/main/scala/ai/starlake/job/ingest/ParquetIngestionJob.scala
@@ -66,7 +66,7 @@ class ParquetIngestionJob(
   val schemaHeaders: List[String] = schema.attributes.map(_.name)
 
   /** Load dataset using spark csv reader and all metadata. Does not infer schema. columns not
-    * defined in the schema are dropped fro the dataset (require datsets with a header)
+    * defined in the schema are dropped from the dataset (require datsets with a header)
     *
     * @return
     *   Spark Dataset
@@ -131,10 +131,10 @@ class ParquetIngestionJob(
       orderedAttributes,
       orderedTypes,
       orderedSparkTypes,
-      settings.comet.privacy.options,
-      settings.comet.cacheStorageLevel,
-      settings.comet.sinkReplayToFile,
-      mergedMetadata.emptyIsNull.getOrElse(settings.comet.emptyIsNull)
+      settings.appConfig.privacy.options,
+      settings.appConfig.cacheStorageLevel,
+      settings.appConfig.sinkReplayToFile,
+      mergedMetadata.emptyIsNull.getOrElse(settings.appConfig.emptyIsNull)
     )
 
     saveRejected(validationResult.errors, validationResult.rejected).map { _ =>