+++ src/test/scala/ai/starlake/job/infer/InferSchemaSpec.scala
@@ -14,7 +14,7 @@ class InferSchemaSpec extends TestHelper {
           |  --table <value>        Table Name
           |  --input <value>        Dataset Input Path
           |  --output-dir <value>   Domain YAML Output Path
-          |  --format <value>       Force format
+          |  --format <value>       Force input file format
           |  --with-header          Does the file contain a header (For CSV files only)
           |""".stripMargin
       rendered.substring(rendered.indexOf("Usage:")).replaceAll("\\s", "") shouldEqual expected