+++ src/test/scala/ai/starlake/schema/generator/Xls2YmlDomainsSpec.scala
@@ -226,7 +226,7 @@ class Xls2YmlDomainsSpec extends TestHelper {
         """
           |Usage: starlake xls2yml [options]
           |
-          |  --files <value>       List of Excel files describing Domains & Schemas OR Jobs
+          |  --files <value>       List of Excel files describing domains & schemas or jobs
           |  --encryption <value>  If true generate pre and post encryption YML
           |  --iamPolicyTagsFile <value>
           |                        If true generate IAM PolicyTags YML
@@ -234,7 +234,7 @@ class Xls2YmlDomainsSpec extends TestHelper {
           |  --privacy <value>     What privacy policies should be applied in the pre-encryption phase ?
           | All privacy policies are applied by default.
           |  --outputPath <value>  Path for saving the resulting YAML file(s).
-          | Comet domains path is used by default.
+          | Starlake domains path is used by default.
           |  --policyFile <value>  Optional File for centralising ACL & RLS definition.
           |  --job <value>         If true generate YML for a Job.
           |""".stripMargin